Correctly select modules/deps based on OS
Now the x11/xdo crates will only be included when on linux. Similar ones
to be created for windows, and later mac os support.
